+++ src/main/java/com/app/OrderItemApplication.java
@@ -1,4 +1,4 @@
-package brijesh.bhatt;
+package com.app;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
